+++ components/calendar/src/cal/buddhist.rs
@@ -17,9 +17,9 @@
 //! ```
 
 use crate::any_calendar::AnyCalendarKind;
+use crate::cal::iso::{Iso, IsoDateInner};
 use crate::calendar_arithmetic::ArithmeticDate;
 use crate::error::DateError;
-use crate::iso::{Iso, IsoDateInner};
 use crate::{types, Calendar, Date, DateDuration, DateDurationUnit, RangeError};
 use tinystr::tinystr;
