+++ components/calendar/src/cal/coptic.rs
@@ -16,9 +16,9 @@
 //! assert_eq!(date_coptic.day_of_month().0, 24);
 //! ```
 
+use crate::cal::iso::Iso;
 use crate::calendar_arithmetic::{ArithmeticDate, CalendarArithmetic};
 use crate::error::DateError;
-use crate::iso::Iso;
 use crate::{types, Calendar, Date, DateDuration, DateDurationUnit, RangeError};
 use calendrical_calculations::helpers::I32CastError;
 use calendrical_calculations::rata_die::RataDie;
