+++ components/calendar/src/cal/gregorian.rs
@@ -16,9 +16,9 @@
 //! assert_eq!(date_gregorian.day_of_month().0, 2);
 //! ```
 
+use crate::cal::iso::{Iso, IsoDateInner};
 use crate::calendar_arithmetic::ArithmeticDate;
 use crate::error::DateError;
-use crate::iso::{Iso, IsoDateInner};
 use crate::{types, Calendar, Date, DateDuration, DateDurationUnit, RangeError};
 use tinystr::tinystr;
