+++ components/calendar/src/cal/indian.rs
@@ -16,9 +16,9 @@
 //! assert_eq!(date_indian.day_of_month().0, 12);
 //! ```
 
+use crate::cal::iso::Iso;
 use crate::calendar_arithmetic::{ArithmeticDate, CalendarArithmetic};
 use crate::error::DateError;
-use crate::iso::Iso;
 use crate::{types, Calendar, Date, DateDuration, DateDurationUnit, RangeError};
 use tinystr::tinystr;
