+++ components/calendar/src/cal/japanese.rs
@@ -24,8 +24,8 @@
 //! );
 //! ```
 
+use crate::cal::iso::{Iso, IsoDateInner};
 use crate::error::DateError;
-use crate::iso::{Iso, IsoDateInner};
 use crate::provider::{CalendarJapaneseExtendedV1, CalendarJapaneseModernV1, EraStartDate};
 use crate::types::Era;
 use crate::{types, AsCalendar, Calendar, Date, DateDuration, DateDurationUnit, Ref};
