+++ components/calendar/src/cal/julian.rs
@@ -17,9 +17,9 @@
 //! ```
 
 use crate::any_calendar::AnyCalendarKind;
+use crate::cal::iso::Iso;
 use crate::calendar_arithmetic::{ArithmeticDate, CalendarArithmetic};
 use crate::error::DateError;
-use crate::iso::Iso;
 use crate::{types, Calendar, Date, DateDuration, DateDurationUnit, RangeError};
 use calendrical_calculations::helpers::I32CastError;
 use calendrical_calculations::rata_die::RataDie;
