+++ components/calendar/src/cal/persian.rs
@@ -15,9 +15,9 @@
 //! assert_eq!(persian_date.day_of_month().0, 11);
 //! ```
 
+use crate::cal::iso::Iso;
 use crate::calendar_arithmetic::{ArithmeticDate, CalendarArithmetic};
 use crate::error::DateError;
-use crate::iso::Iso;
 use crate::{types, Calendar, Date, DateDuration, DateDurationUnit, RangeError};
 use ::tinystr::tinystr;
 use calendrical_calculations::helpers::I32CastError;
