+++ components/calendar/src/cal/roc.rs
@@ -17,7 +17,7 @@
 //! ```
 
 use crate::{
-    calendar_arithmetic::ArithmeticDate, error::DateError, iso::IsoDateInner, types, Calendar,
+    cal::iso::IsoDateInner, calendar_arithmetic::ArithmeticDate, error::DateError, types, Calendar,
     Date, Iso, RangeError,
 };
 use calendrical_calculations::helpers::i64_to_saturated_i32;
@@ -30,13 +30,13 @@ const ROC_ERA_OFFSET: i32 = 1911;
 /// The Republic of China (ROC) Calendar
 ///
 /// The [Republic of China calendar] is a solar calendar used in Taiwan and Penghu, as well as by overseas diaspora from
-/// those locations. Months and days are identical to the [`Gregorian`] calendar, while years are counted
+/// those locations. Months and days are identical to the [`Gregorian`](crate::cal::Gregorian) calendar, while years are counted
 /// with 1912, the year of the establishment of the Republic of China, as year 1 of the ROC/Minguo/民国/民國 era.
 ///
 /// [Republic of China calendar]: https://en.wikipedia.org/wiki/Republic_of_China_calendar
 ///
 /// The Republic of China calendar should not be confused with the Chinese traditional lunar calendar
-/// (see [`Chinese`]).
+/// (see [`Chinese`](crate::cal::Chinese)).
 ///
 /// # Era codes
 ///
@@ -47,9 +47,6 @@ const ROC_ERA_OFFSET: i32 = 1911;
 /// # Month codes
 ///
 /// This calendar supports 12 solar month codes (`"M01" - "M12"`)
-///
-/// [`Chinese`]: crate::chinese::Chinese
-/// [`Gregorian`]: crate::Gregorian
 #[derive(Copy, Clone, Debug, Default)]
 #[allow(clippy::exhaustive_structs)] // this type is stable
 pub struct Roc;
