+++ components/calendar/src/lib.rs
@@ -95,78 +95,11 @@ extern crate alloc;
 // Make sure inherent docs go first
 mod date;
 
-/// Types for individual calendars
-pub mod cal {
-    pub use crate::buddhist::Buddhist;
-    pub use crate::chinese::Chinese;
-    pub use crate::coptic::Coptic;
-    pub use crate::dangi::Dangi;
-    pub use crate::ethiopian::{Ethiopian, EthiopianEraStyle};
-    pub use crate::gregorian::Gregorian;
-    pub use crate::hebrew::Hebrew;
-    pub use crate::indian::Indian;
-    pub use crate::islamic::{
-        IslamicCivil, IslamicObservational, IslamicTabular, IslamicUmmAlQura,
-    };
-    pub use crate::iso::Iso;
-    pub use crate::japanese::{Japanese, JapaneseExtended};
-    pub use crate::julian::Julian;
-    pub use crate::persian::Persian;
-    pub use crate::roc::Roc;
-
-    pub use crate::any_calendar::AnyCalendar;
-
-    /// Scaffolding types: You shouldn't need to use these, they need to be public for the `Calendar` trait impl to work.
-    pub mod scaffold {
-        pub use crate::chinese::ChineseDateInner;
-        pub use crate::coptic::CopticDateInner;
-        pub use crate::dangi::DangiDateInner;
-        pub use crate::ethiopian::EthiopianDateInner;
-        pub use crate::gregorian::GregorianDateInner;
-        pub use crate::hebrew::HebrewDateInner;
-        pub use crate::indian::Indian;
-        pub use crate::islamic::{
-            IslamicCivilDateInner, IslamicDateInner, IslamicTabularDateInner,
-            IslamicUmmAlQuraDateInner,
-        };
-        pub use crate::iso::Iso;
-        pub use crate::japanese::Japanese;
-        pub use crate::julian::JulianDateInner;
-        pub use crate::persian::PersianDateInner;
-        pub use crate::roc::RocDateInner;
-
-        pub use crate::any_calendar::AnyDateInner;
-    }
-}
-
+// Public modules
 pub mod any_calendar;
-mod buddhist;
-mod calendar;
-mod calendar_arithmetic;
-mod chinese;
-mod chinese_based;
-mod coptic;
-mod dangi;
-mod duration;
-mod error;
-mod ethiopian;
-mod gregorian;
-mod hebrew;
-mod indian;
-mod islamic;
-mod iso;
-#[cfg(feature = "ixdtf")]
-mod ixdtf;
-mod japanese;
-mod julian;
-mod persian;
+pub mod cal;
 pub mod provider;
-mod roc;
-#[cfg(test)]
-mod tests;
 pub mod types;
-mod week_of;
-
 pub mod week {
     //! Functions for week-of-month and week-of-year arithmetic.
     use crate::week_of;
@@ -177,19 +110,28 @@ pub mod week {
     pub use week_of::MIN_UNIT_DAYS;
 }
 
+mod calendar;
+mod calendar_arithmetic;
+mod duration;
+mod error;
 #[cfg(feature = "ixdtf")]
-pub use crate::ixdtf::ParseError;
-#[doc(no_inline)]
-pub use any_calendar::{AnyCalendar, AnyCalendarKind, AnyCalendarPreferences};
+mod ixdtf;
+mod week_of;
+
+// Top-level types
 pub use calendar::Calendar;
 pub use date::{AsCalendar, Date, Ref};
 #[doc(hidden)] // unstable
 pub use duration::{DateDuration, DateDurationUnit};
 pub use error::{DateError, RangeError};
+#[cfg(feature = "ixdtf")]
+pub use ixdtf::ParseError;
+
+// Reexports
 #[doc(no_inline)]
-pub use gregorian::Gregorian;
+pub use any_calendar::{AnyCalendar, AnyCalendarKind, AnyCalendarPreferences};
 #[doc(no_inline)]
-pub use iso::Iso;
+pub use cal::{Gregorian, Iso};
 
 /// Locale preferences used by this crate
 pub mod preferences {
@@ -202,3 +144,6 @@ pub mod preferences {
     #[doc = "\n"] // prevent autoformatting
     pub use icu_locale_core::preferences::extensions::unicode::keywords::IslamicCalendarAlgorithm;
 }
+
+#[cfg(test)]
+mod tests;
