+++ components/calendar/src/provider.rs
@@ -119,7 +119,7 @@ pub struct EraStartDate {
 }
 
 /// A data structure containing the necessary era data for constructing a
-/// [`Japanese`](crate::japanese::Japanese) calendar object
+/// [`Japanese`](crate::cal::Japanese) calendar object
 ///
 /// <div class="stab unstable">
 /// 🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways,
